refactor: unify conversion utilities
