+++ apps/client/src/features/viewers/studio/studioClock.utils.ts
@@ -1,4 +1,5 @@
-import { OntimeEvent } from 'ontime-types';
+import { MaybeNumber, OntimeEvent } from 'ontime-types';
+import { MILLIS_PER_MINUTE, MILLIS_PER_SECOND } from 'ontime-utils';
 
 /**
  * @description Returns trimmed event list array
@@ -18,4 +19,16 @@ export function trimRundown(rundown: OntimeEvent[], selectedId: string | null, l
   const endIndex = Math.min(startIndex + limit, rundown.length);
   const trimmedRundown = rundown.slice(startIndex, endIndex);
   return trimmedRundown;
+}
+
+/**
+ * @description Returns amount of seconds in a date given in milliseconds. For studio clock second indicator
+ * @param {MaybeNumber} millis time to format
+ * @returns amount of elapsed seconds
+ */
+export function secondsInMillis(millis: MaybeNumber): number {
+  if (!millis) {
+    return 0;
+  }
+  return Math.floor((millis % MILLIS_PER_MINUTE) / MILLIS_PER_SECOND);
 }
